+++ src/vs/base/node/stream.ts
@@ -5,63 +5,6 @@
 
 import * as fs from 'fs';
 
-export interface ReadResult {
-	buffer: Buffer | null;
-	bytesRead: number;
-}
-
-/**
- * Reads totalBytes from the provided file.
- */
-export function readExactlyByFile(file: string, totalBytes: number): Promise<ReadResult> {
-	return new Promise<ReadResult>((resolve, reject) => {
-		fs.open(file, 'r', null, (err, fd) => {
-			if (err) {
-				return reject(err);
-			}
-
-			function end(err: Error | null, resultBuffer: Buffer | null, bytesRead: number): void {
-				fs.close(fd, closeError => {
-					if (closeError) {
-						return reject(closeError);
-					}
-
-					if (err && (<any>err).code === 'EISDIR') {
-						return reject(err); // we want to bubble this error up (file is actually a folder)
-					}
-
-					return resolve({ buffer: resultBuffer, bytesRead });
-				});
-			}
-
-			const buffer = Buffer.allocUnsafe(totalBytes);
-			let offset = 0;
-
-			function readChunk(): void {
-				fs.read(fd, buffer, offset, totalBytes - offset, null, (err, bytesRead) => {
-					if (err) {
-						return end(err, null, 0);
-					}
-
-					if (bytesRead === 0) {
-						return end(null, buffer, offset);
-					}
-
-					offset += bytesRead;
-
-					if (offset === totalBytes) {
-						return end(null, buffer, offset);
-					}
-
-					return readChunk();
-				});
-			}
-
-			readChunk();
-		});
-	});
-}
-
 /**
  * Reads a file until a matching string is found.
  *
